feat: added dedicated fe/api
